control proxy args through environment variables
We make it possible to control proxy arguments with environment
variables. This makes it easier to modify proxy behavior when running
`yarn start`.

To accomplish this we switch from `argh` to `structopt`.
