+++ proxy/api/Cargo.toml
@@ -18,7 +18,6 @@ unsafe-fast-keystore = []
 
 [dependencies]
 anyhow = "1.0"
-argh = "0.1"
 async-stream = "0.3"
 chrono = { version = "0.4.19", features = [ "serde" ] }
 data-encoding = "2.3"
@@ -37,6 +36,7 @@ serde = { version = "1.0", features = [ "derive" ] }
 serde_json = "1.0"
 serde_qs = "0.8"
 secstr = { version = "0.3.2", features = [ "serde" ] }
+structopt = "0.3.25"
 tempfile = "3.1"
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter", "json"] }
