+++ proxy/api/src/process.rs
@@ -8,51 +8,13 @@
 // Otherwise clippy complains about FromArgs
 #![allow(clippy::default_trait_access)]
 
-use std::{net, sync::Arc, time::Duration};
+use std::{sync::Arc, time::Duration};
 
-use argh::FromArgs;
 use futures::prelude::*;
 use thiserror::Error;
 use tokio::sync::{broadcast, watch, RwLock};
 
-use crate::{config, context, git_helper, http, notification, service, session};
-
-/// Flags accepted by the proxy binary.
-#[derive(Clone, FromArgs)]
-#[allow(clippy::struct_excessive_bools)]
-pub struct Args {
-    /// put proxy in test mode to use certain fixtures
-    #[argh(switch)]
-    pub test: bool,
-    /// run HTTP API on a specified address:port (default: 127.0.0.1:17246)
-    #[argh(
-        option,
-        default = "std::net::SocketAddr::from(([127, 0, 0, 1], 17246))"
-    )]
-    pub http_listen: net::SocketAddr,
-    /// run the peer on a specified address:port (default: 0.0.0.0:0)
-    #[argh(option, default = "std::net::SocketAddr::from(([0, 0, 0, 0], 0))")]
-    pub peer_listen: net::SocketAddr,
-    /// add one or more default seed addresses to initialise the settings store (default: none)
-    #[argh(option, long = "default-seed")]
-    pub default_seeds: Vec<String>,
-    /// don’t install the git-remote-rad binary
-    #[argh(switch)]
-    pub skip_remote_helper_install: bool,
-    #[argh(option)]
-    /// passphrase to unlock the keystore
-    ///
-    /// If not provided the keystore must be unlocked via the HTTP API.
-    pub key_passphrase: Option<String>,
-    #[cfg(feature = "unsafe-fast-keystore")]
-    /// enables fast but unsafe encryption of the keystore for development builds
-    #[argh(switch)]
-    pub unsafe_fast_keystore: bool,
-
-    /// enables more verbose logging for development
-    #[argh(switch)]
-    pub dev_log: bool,
-}
+use crate::{cli::Args, config, context, git_helper, http, notification, service, session};
 
 /// Run the proxy process
 ///
